Update: DrawCanvasクラスのロジックを更新
