+++ index.html
@@ -7,7 +7,7 @@
     <title>Pixeleria Editor</title>
   </head>
   <body>
-    <pixeleria-editor dotsize="64" artwidth="8" artheight="8"></pixeleria-editor>
+    <pixeleria-editor dotsize="4" artwidth="128" artheight="128"></pixeleria-editor>
     <script type="module" src="/src/pixeleria-editor.svelte"></script>
   </body>
 </html>
