+++ src/canvas/canvas-context.ts
