+++ src/canvas/pixel-buffer.ts
