--- src/canvas/preview-canvas.ts
@@ -1,7 +1,0 @@
-import { PixelCanvas } from "./pixel-canvas";
-
-export class PreviewCanvas extends PixelCanvas {
-  constructor(canvas: HTMLCanvasElement) {
-    super(canvas);
-  }
-}
